Mention known pkg.go.dev issue in readme (#11)
Currently, there is an bug with pkg.go.dev with regard to generics.
Examples with generics are not correctly associated to their function
signatures. This causes all the examples to be missing from pkg.go.dev. A
simple workaround is to use `godoc` to read the documentation locally.
Up-to-date local versions of `godoc` renders examples correctly.